+++ .eslintrc.cjs
@@ -1,3 +1,0 @@
-module.exports = {
-    extends: ["hive", "hive/prettier"]
-};