+++ types/.eslintrc.cjs
@@ -1,8 +1,0 @@
-module.exports = {
-    extends: ["hive", "eslint:recommended", "plugin:@typescript-eslint/recommended"],
-    parser: "@typescript-eslint/parser",
-    plugins: ["@typescript-eslint"],
-    rules: {
-        "@typescript-eslint/no-explicit-any": "off"
-    }
-};